feat: 리뷰 업데이트 추가

+++ src/main/java/com/hamtaro/sunflowerplate/service/member/MyPageService.java
@@ -30,10 +30,8 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
+
 @Slf4j
 @Service
 @RequiredArgsConstructor
@@ -80,20 +78,29 @@ public class MyPageService {
         return ResponseEntity.status(200).body(requestMyReviewDtoList);
     }
 
-    public ResponseEntity<?> deleteMyReview(Long reviewId) {
+    public ResponseEntity<?> deleteMyReview(Long reviewId, String userId) {
         Optional<List<ReviewImageEntity>> reviewImageId = reviewImageRepository.findReviewImageEntityByReviewEntity_ReviewId(reviewId);
-        if (reviewImageId.isPresent()) {
-            for (ReviewImageEntity reviewImageEntity : reviewImageId.get()) {
-                reviewService.deleteS3Image(reviewImageEntity.getReviewStoredName());
-                reviewService.deleteS3Image(reviewImageEntity.getReviewResizeStoredName());
+        ReviewEntity findReview = reviewRepository.findById(reviewId).get();
+        if (findReview.getMemberEntity().getMemberId().equals(Long.valueOf(userId))) {
+
+            if (reviewImageId.isPresent()) {
+                for (ReviewImageEntity reviewImageEntity : reviewImageId.get()) {
+                    reviewService.deleteS3Image(reviewImageEntity.getReviewStoredName());
+                    reviewService.deleteS3Image(reviewImageEntity.getReviewResizeStoredName());
+                }
+                reviewRepository.deleteById(reviewId);
+                return ResponseEntity.status(200).body("리뷰삭제");
+            } else {
+                reviewRepository.deleteById(reviewId);
+                return ResponseEntity.status(200).body("리뷰삭제");
             }
-            reviewRepository.deleteById(reviewId);
-            return ResponseEntity.status(200).body("리뷰삭제");
-        } else {
-            reviewRepository.deleteById(reviewId);
-            return ResponseEntity.status(200).body("리뷰삭제");
+        }else{
+            Map<String, String> response = new HashMap<>();
+            response.put("message", "권한없음");
+            return ResponseEntity.status(403).body(response);
         }
     }
+
     private File resizeImage(MultipartFile originalImage) throws IOException {
         File resizeFile = new File("resized_" + originalImage.getOriginalFilename());
         Thumbnails.of(originalImage.getInputStream())
@@ -101,6 +108,7 @@ public class MyPageService {
                   .toFile(resizeFile);
         return resizeFile;
     }
+
     private String getFileExtension(String filename) {
         try {
             return filename.substring(filename.lastIndexOf("."));
@@ -108,6 +116,7 @@ public class MyPageService {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "잘못된 형식의 파일(" + filename + ")입니다.");
         }
     }
+
     private String createFileName(String fileName) {
         return UUID.randomUUID().toString().concat(getFileExtension(fileName));
     }
@@ -115,24 +124,36 @@ public class MyPageService {
     // 업데이트 아직 진행중
     public ResponseEntity<?> updateMyReview(Long reviewId, UpdateReviewDto updateReviewDto) {
         ReviewEntity findReview = reviewRepository.findById(reviewId).get();
-
-        for (ReviewImageEntity reviewImageEntity : findReview.getReviewImageEntityList()) {
-            reviewService.deleteS3Image(reviewImageEntity.getReviewStoredName());
-            reviewService.deleteS3Image(reviewImageEntity.getReviewResizeStoredName());
-        }
-        findReview.setReviewContent(updateReviewDto.getReviewContent());
-        Long saveReviewId = reviewRepository.save(findReview).getReviewId();
-        ReviewEntity updateReview = reviewRepository.findById(saveReviewId).get();
-
-        for (MultipartFile multipartFile: updateReviewDto.getImageFile()){
-
-            String fileName = multipartFile.getOriginalFilename(); //원본 파일
-
-            try {
-                //이미지 객체 생성
-                ObjectMetadata objectMetadata = new ObjectMetadata();
-                objectMetadata.setContentType(multipartFile.getContentType());
-                objectMetadata.setContentLength(multipartFile.getSize());
+        if (findReview.getMemberEntity().getMemberId().equals(Long.valueOf(userId))) {
+            for (UpdateReviewImageDto updateReviewImageDto : updateReviewDto.getImageDtoList()) {
+                for (ReviewImageEntity reviewImageEntity : findReview.getReviewImageEntityList()) {
+                    if (updateReviewImageDto.getImageId().equals(reviewImageEntity.getReviewImageId())) {
+                        reviewService.deleteS3Image(reviewImageEntity.getReviewStoredName());
+                        reviewService.deleteS3Image(reviewImageEntity.getReviewResizeStoredName());
+                        reviewImageRepository.deleteById(reviewImageEntity.getReviewImageId());
+                    }
+                }
+            }
+            List<ReviewImageEntity> reviewImageEntityList = reviewImageRepository.findReviewImageEntityByReviewEntity_ReviewId(reviewId)
+                                                                                 .get();
+            findReview.setReviewImageEntityList(reviewImageEntityList);
+            Long delSaveReviewId = reviewRepository.save(findReview).getReviewId();
+            if (updateReviewDto.getImageDtoList().size() == 3) {
+                for (ReviewImageEntity reviewImageEntity : findReview.getReviewImageEntityList()) {
+                    reviewService.deleteS3Image(reviewImageEntity.getReviewStoredName());
+                    reviewService.deleteS3Image(reviewImageEntity.getReviewResizeStoredName());
+                    reviewImageRepository.deleteById(reviewImageEntity.getReviewImageId());
+                }
+                findReview.setReviewContent(updateReviewDto.getReviewContent());
+                Long saveReviewId = reviewRepository.save(findReview).getReviewId();
+                ReviewEntity updateReview = reviewRepository.findById(saveReviewId).get();
+                for (MultipartFile multipartFile : imageFile) {
+                    String fileName = multipartFile.getOriginalFilename(); //원본 파일
+                    try {
+                        //이미지 객체 생성
+                        ObjectMetadata objectMetadata = new ObjectMetadata();
+                        objectMetadata.setContentType(multipartFile.getContentType());
+                        objectMetadata.setContentLength(multipartFile.getSize());
 
                 //원본 파일 저장
                 String storedName = createFileName(fileName);
@@ -146,31 +167,120 @@ public class MyPageService {
                 amazonS3Client.putObject(bucketName,resizeName,resizedImageFile);
                 String resizeUrl = amazonS3Client.getUrl(bucketName,resizeName).toString();
 
-                //이미지 저장
-                reviewImageRepository.save(ReviewImageEntity.builder()
-                                                            .reviewOriginName(fileName)
-                                                            .reviewStoredName(storedName)
-                                                            .reviewResizeStoredName(resizeName)
-                                                            .reviewOriginUrl(accessUrl)
-                                                            .reviewResizeUrl(resizeUrl)
-                                                            .reviewEntity(updateReview)
-                                                            .build());
+                        //이미지 저장
+                        reviewImageRepository.save(ReviewImageEntity.builder()
+                                                                    .reviewOriginName(fileName)
+                                                                    .reviewStoredName(storedName)
+                                                                    .reviewResizeStoredName(resizeName)
+                                                                    .reviewOriginUrl(accessUrl)
+                                                                    .reviewResizeUrl(resizeUrl)
+                                                                    .reviewEntity(updateReview)
+                                                                    .build());
 
-                if (resizedImageFile != null && resizedImageFile.exists()) {
-                    if (resizedImageFile.delete()) {
-                        log.info("이미지 삭제됨");
-                    } else {
-                        log.info("이미지 삭제");
+                        if (resizedImageFile != null && resizedImageFile.exists()) {
+                            if (resizedImageFile.delete()) {
+                                log.info("이미지 삭제됨");
+                            } else {
+                                log.info("이미지 삭제");
+                            }
+                        } else {
+                            log.info("이미지파일 없음");
+                        }
+                    } catch (IOException e) {
+                        throw new RuntimeException("이미지 업로드에 실패했습니다.");
                     }
-                } else {
-                    log.info("이미지파일 없음");
                 }
-            } catch (IOException e){
-                throw new RuntimeException("이미지 업로드에 실패했습니다.");
+            } else if (updateReviewDto.getImageDtoList().size() > 0) {
+                ReviewEntity updateReviewEntity = reviewRepository.findById(delSaveReviewId).get();
+                updateReviewEntity.setReviewContent(updateReviewDto.getReviewContent());
+                Long saveReviewId = reviewRepository.save(updateReviewEntity).getReviewId();
+                ReviewEntity updateReview = reviewRepository.findById(saveReviewId).get();
+                for (MultipartFile multipartFile : imageFile) {
+
+                    String fileName = multipartFile.getOriginalFilename(); //원본 파일
+
+                    try {
+                        //이미지 객체 생성
+                        ObjectMetadata objectMetadata = new ObjectMetadata();
+                        objectMetadata.setContentType(multipartFile.getContentType());
+                        objectMetadata.setContentLength(multipartFile.getSize());
+
+                        //원본 파일 저장
+                        String storedName = createFileName(fileName);
+                        amazonS3Client.putObject(new PutObjectRequest(bucketName, storedName, multipartFile.getInputStream(), objectMetadata));
+                        String accessUrl = amazonS3Client.getUrl(bucketName, storedName).toString();
+                        System.out.println("Original Image URL:" + accessUrl);
+
+                        //리사이징 파일 저장
+                        String resizeName = "resized_" + storedName;
+                        File resizedImageFile = resizeImage(multipartFile);
+                        amazonS3Client.putObject(bucketName, resizeName, resizedImageFile);
+                        String resizeUrl = amazonS3Client.getUrl(bucketName, resizeName).toString();
+                        ReviewImageEntity reviewImageEntity = ReviewImageEntity.builder()
+                                                                               .reviewOriginName(fileName)
+                                                                               .reviewStoredName(storedName)
+                                                                               .reviewResizeStoredName(resizeName)
+                                                                               .reviewOriginUrl(accessUrl)
+                                                                               .reviewResizeUrl(resizeUrl)
+                                                                               .reviewEntity(updateReview)
+                                                                               .build();
+                        //이미지 저장
+                        reviewImageRepository.save(reviewImageEntity);
+
+                        if (resizedImageFile != null && resizedImageFile.exists()) {
+                            if (resizedImageFile.delete()) {
+                                log.info("이미지 삭제됨");
+                            } else {
+                                log.info("이미지 삭제");
+                            }
+                        } else {
+                            log.info("이미지파일 없음");
+                        }
+                    } catch (IOException e) {
+                        throw new RuntimeException("이미지 업로드에 실패했습니다.");
+                    }
+                }
+            } else {
+                ReviewEntity updateReviewEntity = reviewRepository.findById(delSaveReviewId).get();
+                updateReviewEntity.setReviewContent(updateReviewDto.getReviewContent());
+                reviewRepository.save(updateReviewEntity);
             }
+            return true;
+        } else {
+
+            return false;
         }
-        return null;
     }
+
+    public ResponseEntity<?> updateMyReview(Long reviewId) {
+        ReviewEntity responseEntity = reviewRepository.findById(reviewId).get();
+        List<ReviewImageDto> reviewImageDtoList = new ArrayList<>();
+
+        for (ReviewImageEntity reviewImageEntity : reviewImageRepository.findReviewImageEntityByReviewEntity_ReviewId(reviewId).get()) {
+            ReviewImageDto reviewImageDto =
+                    ReviewImageDto.builder()
+                                  .reviewImageId(reviewImageEntity.getReviewImageId())
+                                  .reviewOriginName(reviewImageEntity.getReviewOriginName())
+                                  .reviewOriginUrl(reviewImageEntity.getReviewOriginUrl())
+                                  .reviewResizeUrl(reviewImageEntity.getReviewResizeUrl())
+                                  .reviewStoredName(reviewImageEntity.getReviewStoredName())
+                                  .build();
+            reviewImageDtoList.add(reviewImageDto);
+        }
+
+        RequestMyReviewDto requestMyReviewDto =
+                RequestMyReviewDto.builder()
+                                  .restaurantId(responseEntity.getRestaurantEntity().getRestaurantId()) // 레스토랑ID
+                                  .restaurantName(responseEntity.getRestaurantEntity().getRestaurantName()) //레스토랑이름
+                                  .reviewContent(responseEntity.getReviewContent()) //리뷰내용
+                                  .reviewStarRating(responseEntity.getReviewStarRating()) //리뷰별점
+                                  .reviewImageDto(reviewImageDtoList)
+                                  .reviewAt(responseEntity.getReviewAt()) // 리뷰작성시간
+                                  .build();
+        return ResponseEntity.status(200).body(requestMyReviewDto);
+
+    }
+
 
     public ResponseEntity<?> getMyPlace(String userId) {
         MemberEntity findId = memberRepository.findById(Long.valueOf(userId)).get();
